changed apis around
