+++ controllers/index.js
@@ -1,9 +1,9 @@
 const router = require('express').Router();
 const apiRoutes = require('./api');
-const homeRoutes = require('./homeRoutes');
+// const homeRoutes = require('./homeRoutes');
 const authController = require('./authController')
 
-router.use('/', homeRoutes);
+// router.use('/', homeRoutes);
 router.use('/api', apiRoutes);
 router.use('/', authController);
